refactor(test): extract findByTag integration tests into own spec
- move findByTag context to PostgresVersionRepositoryFindByTagIntegrationSpec
- keeps PostgresVersionRepositoryIntegrationSpec under detekt LargeClass limit
